+++ examples/04_geo_h2/run_geo_h2.py
@@ -12,12 +12,16 @@ import matplotlib.pyplot as plt
 
 
 hydrogen_out = h2i_nat.prob.model.plant.geoh2_well_subsurface.NaturalGeoH2PerformanceModel.get_val(
-    "hydrogen_out", units="kg/h"
+    "annual_hydrogen_produced", units="t/year"
 )
-plt.plot(hydrogen_out)
-plt.xlabel("Time (hours)")
-plt.ylabel("Wellhead Gas Flow (kg/h)")
-plt.title("Wellhead Gas Flow Profile Over First Year")
+start_year = (
+    h2i_nat.prob.model.plant.finance_subgroup_h2.hydrogen_finance_default.params.analysis_start_year
+)
+years = range(start_year, start_year + len(hydrogen_out))
+plt.plot(years, hydrogen_out)
+plt.xlabel("Year")
+plt.ylabel("Hydrogen Production (tonne per annum)")
+plt.title("Hydrogen Production over Well Lifetime")
 plt.grid()
 plt.show()